# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Esta é a API de Indicadores.",
        "title": "API de Indicadores",
        "contact": {
            "name": "Suporte",
            "email": "analytics@controllab.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://victor.controllab.com:8000/api/v1"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return the logged-in user's data",
                "tags": [
                    "user"
                ],
                "summary": "User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "members": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/variable": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all variables",
                "tags": [
                    "variables"
                ],
                "summary": "List variables",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Variavel"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create only one variable",
                "tags": [
                    "variables"
                ],
                "summary": "Create variable",
                "parameters": [
                    {
                        "description": "Variable",
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Variavel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/createResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/variable/groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all groups by user",
                "tags": [
                    "variables"
                ],
                "summary": "List Groups by user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Grupamento"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "variables"
                ],
                "summary": "Create groups",
                "parameters": [
                    {
                        "description": "Groups",
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Grupamento"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/createResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/variable/tipo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all variables types",
                "tags": [
                    "variables"
                ],
                "summary": "List types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TipoVariavel"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "createResponse": {
                "type": "object",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "Empresa": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "level": {
                        "$ref": "#/components/schemas/Nivel"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
...
            "Grupamento": {
                "type": "object",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
...
            "Nivel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string"
                    }
                }
            },
...
            "TipoVariavel": {
                "type": "object",
                "required": [
                    "nome"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string"
                    }
                }
            },
            "User": {
                "type": "object",
                "required": [
                    "first_name",
                    "last_name",
                    "password",
                    "user"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "enterprise": {
                        "$ref": "#/components/schemas/Empresa"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "groups": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Grupamento"
                        }
                    },
                    "id": {
                        "type": "integer"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "manager_id": {
                        "type": "integer"
                    },
                    "password": {
                        "type": "string"
                    },
                    "user": {
                        "type": "string"
                    }
                }
            },
            "Variavel": {
                "type": "object",
                "required": [
                    "data_type_id",
                    "text"
                ],
                "properties": {
                    "data_type_id": {
                        "$ref": "#/components/schemas/TipoVariavel"
                    },
                    "has_item": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "question_id": {
                        "type": "integer"
                    },
                    "required": {
                        "type": "boolean"
                    },
                    "text": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}